Added server with CRUD capability.
This is a server using express. Express is the most popular way to right rest api's in javascript. I tried to keep things fairly clean and simple. The server does currently support all CRUD functionality as well as pagination. I still need to add search to it. The biggest ommision though is documentation. I plan on adding documentation around running the code, more comments, and also "swagger" documentation which will show the endpoints and how to use them. I will also enhance the paging with links in headers.

The only change I did to existing code is upgrading the react-scripts to 2.3. This allows us to using proxying for the backend. It greatly simplifies setup for this.

If you have any questions let me know.